+++ core/src/test/java/com/homeaway/streamplatform/streamregistry/health/StreamRegistryHealthCheckIT.java
@@ -22,6 +22,7 @@ import org.junit.Test;
 
 import com.homeaway.streamplatform.streamregistry.resource.BaseResourceIT;
 
+
 /**
  * Test class to validate whether the StreamRegistryHealthCheck could properly insert
  * a valid Stream, producer and consumer, and validate whether the StateStore is active
@@ -31,11 +32,12 @@ public class StreamRegistryHealthCheckIT extends BaseResourceIT {
 
     @Test
     public void testHealthCheck() {
-        // The identityMapCapacity on CachedSchemaRegistryClient is set to 100.
-        // so, hit healthcheck (update same schema) for more than 100 times to make sure the applied fix for issue#100 is working.
-        for (int i=0 ; i < 150 ; i++) {
+        // The identityMapCapacity on CachedSchemaRegistryClient is set to 1 in BaseResourceIT.
+        // so, hit healthcheck (update same schema) for more than 1 times to make sure the applied fix for issue#100 is working.
+        for (int i=0 ; i < 2 ; i++) {
             HealthCheck.Result healthCheckResult = healthCheck.check();
             Assert.assertTrue(healthCheckResult.isHealthy());
         }
     }
+
 }