+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.4.5] - SNAPSHOT
 ### Changed
 - Hardening check for ensuring KStream store is available during Integration Tests
+- Populated the HealthCheck stream region from config file. Removed the dependency of MPAAS_REGION variable. (#103)
 
 ## [0.4.4] - 20190204
 ### Changed